# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentação da API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://sistema-generico-api.vercel.app/api/api"
        }
    ],
    "paths": {
...
        "/estoque/{id_estoque}": {
            "get": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Obtém um estoque pelo ID",
                "operationId": "276a5b0c1316c3ba9337fa2ad35fcb43",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Estoque"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Estoque não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Atualiza um estoque",
                "operationId": "dc56ceaae9eb25a035c2bfa4ed8aa4f7",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Estoque"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estoque atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Deleta um estoque",
                "operationId": "6285ba21235806c605c51620eb5c0c7e",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque deletado com sucesso"
                    }
                }
            }
        },
...
        "/estoqueItem/{id_estoque_item}": {
            "get": {
                "tags": [
                    "EstoqueItem"
                ],
                "summary": "Obtém um estoque item pelo ID",
                "operationId": "72f4dd8e594ffcae28acf4692b2970ee",
                "parameters": [
                    {
                        "name": "id_estoque_item",
                        "in": "path",
                        "description": "ID do Estoque Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque Item encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstoqueItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Estoque Item não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Atualiza um estoque",
                "operationId": "79fe3a55c6941ff7ebf2479150c599c1",
                "parameters": [
                    {
                        "name": "id_estoque_item",
                        "in": "path",
                        "description": "ID do estoque item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Estoque"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estoque atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/estoqueItem": {
            "post": {
                "tags": [
                    "EstoqueItem"
                ],
                "summary": "Cria um novo estoque para o produto enviado",
                "operationId": "c798a58d565c96d5869a8014121fd014",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstoqueItem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estoque criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstoqueItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/estoque/{id_estoque_item}": {
            "delete": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Deleta um estoque daquele produto",
                "operationId": "260ea3b68e28a2445ce8e431d20dfb62",
                "parameters": [
                    {
                        "name": "id_estoque_item",
                        "in": "path",
                        "description": "ID do estoque item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque Item deletado com sucesso"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Estoque": {
                "properties": {
                    "id_estoque_est": {
                        "type": "integer"
                    },
                    "id_empresa_est": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_estoque_est": {
                        "type": "string"
                    },
                    "id_centro_custo_est": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    },
                    "is_ativo_est": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EstoqueItem": {
                "properties": {
                    "id_estoque_item_eti": {
                        "type": "integer"
                    },
                    "id_material_eti": {
                        "description": "Chave estrangeira referenciando o material",
                        "type": "integer"
                    },
                    "id_empresa_eti": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "id_estoque_eti": {
                        "description": "Chave estrangeira referenciando o estoque",
                        "type": "integer"
                    },
                    "des_estoque_item_eti": {
                        "description": "Descrição do item no estoque",
                        "type": "string",
                        "nullable": true
                    },
                    "qtd_estoque_item_eti": {
                        "description": "Quantidade do item no estoque",
                        "type": "number",
                        "format": "float"
                    },
                    "is_ativo_eti": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}